refactored tags property and serialize to baseviewmodel
