+++ GTBA/GTBA/ViewModels/Franchises/EditFranchiseViewModel.cs
@@ -11,12 +11,10 @@ namespace GTBA.ViewModels.Franchises
     public class EditFranchiseViewModel : BaseViewModel
     {
         public Franchise Franchise { get; set; }
-        public ObservableCollection<string> Tags { get; set; }
         public EditFranchiseViewModel(Franchise franchise)
         {
             Franchise = franchise;
             Title = franchise?.FranchiseName;
-            Tags = new ObservableCollection<string>();
             DeserializeTags();
         }
 
@@ -29,21 +27,6 @@ namespace GTBA.ViewModels.Franchises
         public void AddTag(string tag)
         {
             Tags.Add(tag);
-        }
-
-        public string SerializeTags()
-        {
-            String str = "";
-            for (int i = 0; i < Tags.Count; i++)
-            {
-                str += Tags[i];
-                // Do not append comma at the end of last element
-                if (i < Tags.Count - 1)
-                {
-                    str += "#";
-                }
-            }
-            return str;
         }
 
         public void DeserializeTags()
